+++ controllers/PersonController.php
@@ -7,26 +7,46 @@ use yii\web\Controller;
 use app\models\CoachModel;
 use app\models\Person;
 
-class PersonController extends BaseController {
+class PersonController extends BaseController
+{
 
     public $layout = 'inner';
 
-    public function actionIndex() {
+    public function actionIndex()
+    {
+        if (Yii::$app->user->isGuest) {
+            return $this->goHome();
+        }
+
         $persons = Person::browse();
         return $this->render('index', [
                     'persons' => $persons,
         ]);
     }
 
-    public function actionView($id) {
+    public function actionView($id)
+    {
+        if (Yii::$app->user->isGuest) {
+            return $this->goHome();
+        }
+
         $person = Person::findOne(['id' => $id]);
+
+        if (!$person || $person->coach_id != Yii::$app->user->id) {
+            throw new \yii\web\ForbiddenHttpException(Yii::t('app', 'Your not allowed to access this page.'));
+        }
         Yii::$app->session->set('person_id', $id);
         return $this->render('view', [
                     'person' => $person,
         ]);
     }
 
-    public function actionNew() {
+    public function actionNew()
+    {
+        if (Yii::$app->user->isGuest) {
+            return $this->goHome();
+        }
+
         $person = new Person();
 
         if ($person->load(Yii::$app->request->post()) && $person->save()) {
@@ -41,14 +61,23 @@ class PersonController extends BaseController {
         ]);
     }
 
-    public function actionEdit($id) {
+    public function actionEdit($id)
+    {
+        if (Yii::$app->user->isGuest) {
+            return $this->goHome();
+        }
+
         $person = Person::findOne(['id' => $id]);
+
+        if (!$person || $person->coach_id != Yii::$app->user->id) {
+            throw new \yii\web\ForbiddenHttpException(Yii::t('app', 'Your not allowed to access this page.'));
+        }
 
         if ($person->load(Yii::$app->request->post()) && $person->save()) {
             SiteController::addFlash('success', Yii::t('app', '{name} has been successfully edited.', ['name' => $person->fullname]));
             return $this->redirect(['/person']);
         } else {
-           SiteController::FlashErrors($person);
+            SiteController::FlashErrors($person);
         }
 
         return $this->render('form', [
@@ -56,8 +85,18 @@ class PersonController extends BaseController {
         ]);
     }
 
-    public function actionDelete($id) {
+    public function actionDelete($id)
+    {
+        if (Yii::$app->user->isGuest) {
+            return $this->goHome();
+        }
+
         $person = Person::findOne(['id' => $id]);
+
+        if (!$person || $person->coach_id != Yii::$app->user->id) {
+            throw new \yii\web\ForbiddenHttpException(Yii::t('app', 'Your not allowed to access this page.'));
+        }
+
         if ($person->delete()) {
             SiteController::addFlash('success', Yii::t('app', '{name} has been successfully deleted.', ['name' => $person->fullname]));
             return $this->redirect(['/person']);
